requirements.txt
AccessControl==3.0.13
Acquisition==4.2.2
Babel==1.3
Chameleon==2.25
DateTime==4.0.1
DocumentTemplate==2.13.4
ExtensionClass==4.1.2
Jinja2==2.8
Markdown==2.6.6
MarkupSafe==1.0
Missing==2.13.1
MultiMapping==2.13.0
Persistence==2.13.2
Plone==5.0.8
Products.ATContentTypes==2.2.13
Products.Archetypes==1.10.18
Products.BTreeFolder2==2.14.0
Products.CMFCore==2.2.10
Products.CMFDiffTool==3.1.4
Products.CMFDynamicViewFTI==4.1.5
Products.CMFEditions==2.2.23
Products.CMFFormController==3.0.8
Products.CMFPlacefulWorkflow==1.6.7
Products.CMFPlone==5.0.8
Products.CMFQuickInstallerTool==3.0.15
Products.CMFUid==2.2.2
Products.DCWorkflow==2.2.5
Products.DateRecurringIndex==2.1
Products.DocFinderTab==1.0.5
Products.ExtendedPathIndex==3.1.1
Products.ExternalEditor==1.1.3
Products.ExternalMethod==2.13.1
Products.GenericSetup==1.8.8
Products.MIMETools==2.13.0
Products.MailHost==2.13.2
Products.MimetypesRegistry==2.0.10
Products.OFSP==2.13.2
Products.PasswordResetTool==2.2.3
Products.PlacelessTranslationService==2.0.7
Products.PlonePAS==5.0.14
Products.PluggableAuthService==1.11.0
Products.PluginRegistry==1.4
Products.PortalTransforms==2.2.1
Products.PrintingMailHost==1.0
Products.PythonScripts==2.13.2
Products.ResourceRegistries==3.0.5
Products.Sessions==3.0
Products.StandardCacheManagers==2.13.1
Products.TemporaryFolder==3.0
Products.ZCTextIndex==2.13.5
Products.ZCatalog==3.0.2
Products.ZSQLMethods==2.13.5
Products.ZopeVersionControl==1.1.3
Products.contentmigration==2.1.16
Products.statusmessages==4.1.2
Products.validation==2.0.2
Pygments==2.0.2
Record==2.13.0
RestrictedPython==3.6.0
Sphinx==1.3.6
Unidecode==0.4.1
ZConfig==2.9.3
ZODB3==3.10.7
ZServer==3.0
Zope2==2.13.26
ZopeUndo==2.12.0
alabaster==0.7.10
archetypes.multilingual==3.0.5
archetypes.schemaextender==2.1.7
args==0.1.0
borg.localrole==3.1.3
check-manifest==0.36
clint==0.5.1
collective.elephantvocabulary==0.2.5
collective.monkeypatcher==1.1.2
colorama==0.3.9
configparser==3.5.0
cssmin==0.2.0
cssselect==0.9.2
diazo==1.2.5
enum34==1.1.6
feedparser==5.2.1
five.customerize==1.1
five.globalrequest==1.0
five.intid==1.1.2
five.localsitemanager==2.0.6
five.pt==2.2.4
flake8==3.5.0
flake8-blind-except==0.1.1
flake8-coding==1.3.0
flake8-commas==1.0.0
flake8-debugger==1.4.0
flake8-deprecated==1.3
flake8-isort==2.3
flake8-pep3101==1.2
flake8-plone-api==1.4
flake8-plone-hasattr==0.2.post0
flake8-polyfill==1.0.2
flake8-print==3.0.1
flake8-quotes==0.13.0
flake8-string-format==0.2.3
flake8-todo==0.7
future==0.15.2
gocept.zestreleaser.customupload==3.0
i18ndude==4.0.1
icalendar==3.11.4
initgroups==2.13.0
isort==4.2.15
mailinglogger==3.8.0
mccabe==0.6.1
mechanize==0.2.5
mockup==2.1.7
nt-svcutils==2.13.0
ordereddict==1.1
pkginfo==1.4.1
plone.alterego==1.1.1
plone.api==1.7
plone.app.blob==1.7.2
plone.app.caching==1.2.18
plone.app.collection==1.1.7
plone.app.content==3.4.1
plone.app.contentlisting==1.2.9
plone.app.contentmenu==2.1.10
plone.app.contentrules==4.0.16
plone.app.contenttypes==1.2.23
plone.app.controlpanel==3.0.4
plone.app.customerize==1.3.7
plone.app.dexterity==2.3.7
plone.app.discussion==2.4.21
plone.app.event==2.0.13
plone.app.folder==1.2.3
plone.app.i18n==3.0.4
plone.app.imaging==2.0.6
plone.app.intid==1.1.2
plone.app.iterate==3.2.4
plone.app.layout==2.5.23
plone.app.linkintegrity==3.1
plone.app.locales==5.0.14
plone.app.lockingbehavior==1.0.5
plone.app.multilingual==5.0.7
plone.app.openid==2.1.1
plone.app.portlets==3.1.5
plone.app.querystring==1.3.17
plone.app.redirector==1.3.4
plone.app.registry==1.5
plone.app.relationfield==1.3.4
plone.app.robotframework==1.1
plone.app.testing==5.0.6
plone.app.textfield==1.2.8
plone.app.theming==1.3.6
plone.app.upgrade==2.0.4
plone.app.users==2.3.8
plone.app.uuid==1.1.3
plone.app.versioningbehavior==1.3
plone.app.viewletmanager==2.0.10
plone.app.vocabularies==4.0.3
plone.app.widgets==2.0.7
plone.app.workflow==2.2.8
plone.app.z3cform==1.2.3
plone.autoform==1.6.2
plone.batching==1.1.1
plone.behavior==1.2.0
plone.browserlayer==2.2.0
plone.cachepurging==1.0.13
plone.caching==1.1.2
plone.contentrules==2.0.6
plone.dexterity==2.4.5
plone.event==1.3.3
plone.folder==1.0.9
plone.formwidget.namedfile==1.0.15
plone.formwidget.recurrence==2.1.1
plone.i18n==3.0.7
plone.indexer==1.0.4
plone.intelligenttext==2.1.0
plone.keyring==3.0.1
plone.locking==2.1.3
plone.memoize==1.2.0
plone.namedfile==3.0.11
plone.openid==2.0.4
plone.outputfilters==3.0.1
plone.portlet.collection==3.1
plone.portlet.static==3.1
plone.portlets==2.2.3
plone.protect==3.0.23
plone.recipe.codeanalysis==2.3
plone.recipe.zope2instance==4.2.22
plone.registry==1.0.5
plone.resource==1.2.1
plone.resourceeditor==2.0.5
plone.rfc822==1.1.3
plone.scale==1.4.2
plone.schema==1.0.0
plone.schemaeditor==2.0.17
plone.session==3.6.1
plone.stringinterp==1.2.1
plone.subrequest==1.8
plone.supermodel==1.3.3
plone.synchronize==1.0.2
plone.testing==4.1.2
plone.theme==3.0.2
plone.transformchain==1.2.0
plone.uuid==1.0.4
plone.z3cform==0.8.1
plonetheme.barceloneta==1.6.23
ply==3.4
pycodestyle==2.3.1
pyflakes==1.6.0
python-dateutil==2.4.2
python-gettext==3.0
python-openid==2.2.5
pytz==2016.10
repoze.xmliter==0.6
requests==2.9.2
requests-toolbelt==0.8.0
robotframework-debuglibrary==0.5
robotframework-selenium2screenshots==0.7.0
robotsuite==1.7.0
roman==1.4.0
sauna.reload==0.5.5
setuptools==38.2.5
six==1.10.0
slimit==0.8.1
snowballstemmer==1.2.1
sourcecodegen==0.6.14
sphinx-rtd-theme==0.1.9
sphinxcontrib-robotframework==0.7.0
tempstorage==2.12.2
testfixtures==5.3.1
transaction==1.1.1
twine==1.8.1
unittest2==0.5.1
venusian==1.1.0
venusianconfiguration==1.1.0
z3c.autoinclude==0.3.7
z3c.caching==2.0a1
z3c.form==3.2.11
z3c.formwidget.query==0.13
z3c.objpath==1.1
z3c.pt==3.0.0a1
z3c.relationfield==0.7
z3c.zcmlhook==1.0b1
zExceptions==2.13.0
zLOG==2.11.2
zc.lockfile==1.0.2
zc.recipe.egg==2.0.5
zc.recipe.testrunner==1.2.1
zc.relation==1.0
zdaemon==2.0.7
zest.pocompile==1.4
zest.releaser==6.9
zope.annotation==3.5.0
zope.app.file==3.6.1
zope.app.locales==3.6.2
zope.app.publication==3.12.0
zope.authentication==3.7.1
zope.broken==3.6.0
zope.browser==1.3
zope.browsermenu==3.9.1
zope.browserpage==3.12.2
zope.browserresource==3.10.3
zope.cachedescriptors==3.5.1
zope.component==3.9.5
zope.componentvocabulary==1.0.1
zope.configuration==3.7.4
zope.container==3.11.2
zope.contentprovider==3.7.2
zope.contenttype==3.5.5
zope.copy==3.5.0
zope.datetime==3.4.1
zope.deferredimport==3.5.3
zope.deprecation==3.4.1
zope.dottedname==3.4.6
zope.dublincore==3.7.0
zope.error==3.7.4
zope.event==3.5.2
zope.exceptions==3.6.2
zope.filerepresentation==3.6.1
zope.formlib==4.0.6
zope.globalrequest==1.2
zope.i18n==3.7.4
zope.i18nmessageid==3.5.3
zope.interface==3.6.7
zope.intid==3.7.2
zope.keyreference==3.6.4
zope.lifecycleevent==3.6.2
zope.location==3.9.1
zope.pagetemplate==3.6.3
zope.processlifetime==1.0
zope.proxy==3.6.1
zope.ptresource==3.9.0
zope.publisher==3.12.6
zope.ramcache==1.0
zope.schema==4.4.2
zope.security==3.7.4
zope.sendmail==3.7.5
zope.sequencesort==3.4.0
zope.site==3.9.2
zope.size==3.4.1
zope.structuredtext==3.5.1
zope.tal==3.5.2
zope.tales==3.5.3
zope.testbrowser==3.11.1
zope.testing==3.9.7
zope.testrunner==4.4.4
zope.traversing==3.13.2
zope.viewlet==3.7.2
